+++ src/components/EitherListCard.js
@@ -106,8 +106,13 @@ const EitherListCard = props => {
   };
   //삭제하기
   const onClickDelete = () => {
-    dispatch(deletePostDB(eitherId));
-    setAction(true);
+    const deleteConfirm = window.confirm("투표를 삭제하시겠습니까?");
+    if (deleteConfirm === true) {
+      dispatch(deletePostDB(eitherId));
+      setAction(true);
+    } else if (deleteConfirm === false) {
+      return;
+    }
   };
   //좋아요
   const onClickLike = () => {
@@ -134,12 +139,15 @@ const EitherListCard = props => {
   };
   //투표 종료하기
   const onClickComplete = () => {
+    const deleteConfirm = window.confirm("투표를 종료하시겠습니까?");
     if (completed === 1) {
       alert("이미 투표가 종료되었습니다.");
       return;
-    } else {
+    } else if (deleteConfirm === true) {
       dispatch(completePostDB(eitherId));
       setAction(true);
+    } else {
+      return;
     }
   };
   //버튼A 상태 보여주기
@@ -236,9 +244,9 @@ const EitherListCard = props => {
                   <Nickname
                     nickname={nickname}
                     userId={user}
-                    width={"32px"}
-                    height={"32px"}
-                    fontSize={"14px"}
+                    width={"16px"}
+                    height={"16px"}
+                    fontSize={"12px"}
                   />
                 </div>
                 <div className="Grid">
@@ -246,8 +254,8 @@ const EitherListCard = props => {
                   <TotalCntGrid>
                     <FaRegUser
                       style={{
-                        width: "20px",
-                        height: "20px",
+                        width: "16px",
+                        height: "16px",
                         color: "#00397c",
                       }}
                     />
@@ -257,16 +265,16 @@ const EitherListCard = props => {
                     <FiThumbsUp
                       onClick={onClickLike}
                       style={{
-                        width: "24px",
-                        height: "24px",
+                        width: "16px",
+                        height: "16px",
                         cursor: "pointer",
                       }}
                     />
                   ) : (
                     <HiThumbUp
                       style={{
-                        width: "24px",
-                        height: "24px",
+                        width: "16px",
+                        height: "16px",
                         cursor: "pointer",
                       }}
                     />
@@ -339,20 +347,19 @@ const EitherListCard = props => {
 const Container = styled.div`
   max-width: 840px;
   width: 100%;
-  min-height: 258px;
+  min-height: 248px;
   height: 100%;
   box-sizing: border-box;
-  margin: 20px auto;
+  margin: 24px auto;
   border: 2px solid #00397c;
   border-radius: 10px;
   background-color: #ffffff;
   padding: 44px 70px 60px 56px;
   position: relative;
   @media screen and (max-width: ${mobile}) {
-    transform: scale(80%);
-    margin: -20px auto;
+    margin: 30px auto;
     padding: 10% 6%;
-    min-height: 240px;
+    width: 80%;
   }
 `;
 const MenuButtonGrid = styled.div`
@@ -388,24 +395,25 @@ const RightGrid = styled.div`
 `;
 const TitleDiv = styled.div`
   width: 100%;
-  font-size: 20px;
+  font-size: 18px;
   font-weight: bold;
   word-break: break-all;
-  @media screen and (max-width: ${mobile}) {
-    font-size: 17px;
-  }
 `;
 const DateDiv = styled.div`
   margin: 10px 0px;
-  font-size: 11px;
+  font-size: 14px;
   color: #868e96;
   position: relative;
+  @media screen and (max-width: ${mobile}) {
+    font-size: 12px;
+  }
 `;
 const ButtonGrid = styled.div`
   max-width: 365px;
   width: 100%;
   height: 100%;
   margin: 0px;
+  display: inline-flex;
 `;
 const EitherButtonA = styled.button`
   width: 50%;
@@ -414,8 +422,6 @@ const EitherButtonA = styled.button`
   border: 2px solid #00397c;
   border-top-left-radius: 10px;
   border-bottom-left-radius: 10px;
-  font-size: 16px;
-  line-height: 23px;
   cursor: pointer;
   margin-right: -1px;
   @media screen and (max-width: ${mobile}) {
@@ -429,8 +435,6 @@ const EitherButtonB = styled.button`
   border: 2px solid #00397c;
   border-top-right-radius: 10px;
   border-bottom-right-radius: 10px;
-  font-size: 16px;
-  line-height: 23px;
   cursor: pointer;
   margin-left: -1px;
   @media screen and (max-width: ${mobile}) {
@@ -438,12 +442,14 @@ const EitherButtonB = styled.button`
   }
 `;
 const ButtonText = styled.div`
+  max-width: 65px;
+  width: 100%;
   word-break: break-all;
   display: inline-block;
   vertical-align: middle;
   font-size: 13px;
   font-weight: bold;
-  padding: 5px 3px;
+  font-family: "Noto-Sans KR", sans-serif;
 `;
 const ProgressGrid = styled.div`
   width: 100%;
@@ -489,35 +495,35 @@ const EitherFooter = styled.div`
   left: 56px;
   align-items: center;
   @media screen and (max-width: ${mobile}) {
-    top: 75.5%;
-    left: 5%;
+    top: 67%;
+    left: 8%;
   }
-
   .Position {
     width: 100%;
     display: flex;
     justify-content: space-between;
     align-items: center;
+    @media screen and (max-width: ${mobile}) {
+      flex-direction: column;
+      justify-content: flex-start;
+    }
   }
   .Profile {
     width: 100%;
     @media screen and (max-width: ${mobile}) {
-      transform: scale(0.9);
+      margin-bottom: 10px;
     }
   }
   .Grid {
+    width: 100%;
     color: #e25b45;
     display: flex;
     align-items: center;
-    @media screen and (max-width: ${mobile}) {
-      transform: scale(0.8);
-    }
   }
   .Likes {
     font-size: 14px;
     margin-left: 14px;
     @media screen and (max-width: ${mobile}) {
-      transform: scale(0.8);
       margin-left: 10px;
     }
   }
@@ -528,14 +534,13 @@ const TotalCntGrid = styled.div`
   align-items: center;
   margin-right: 20px;
   @media screen and (max-width: ${mobile}) {
-    transform: scale(0.9);
     margin-right: 10px;
   }
 `;
 const TotalCntDiv = styled.div`
   font-size: 14px;
   color: #868e96;
-  margin-left: 8px;
+  margin-left: 10px;
 `;
 
 export default EitherListCard;
